backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
Pillow>=10.0.0
zai-sdk>=0.0.3
httpx>=0.24.0
python-multipart>=0.0.5
python-dotenv>=0.19.0